Extract filename parsing into testable class
